+++ src/backend/adapters/WishListAdapter.test.ts
@@ -5,7 +5,7 @@ import path from "path"
 import { afterAll, beforeAll, describe, expect, test } from "vitest"
 import { DbWishItem } from "../domain/models/WishList"
 import { runMigrations } from "../utils/migrationManager"
-import { sql } from "../utils/sql.old"
+import { sql } from "../utils/sql"
 import { makeSqliteConnection } from "../utils/sqliteConnection"
 import { makeSqliteUserRepository } from "./SqliteUserRepository"
 import { makeWishListStorageAdapter } from "./WishListAdapter"
